[NNAdapter][HuaweiAscendNPU] Fix the compilation error caused by the introduction of ACL GE header file for Linux+x86 test=develop

+++ lite/backends/nnadapter/nnadapter/driver/huawei_ascend_npu/converter/fully_connected.cc
@@ -64,7 +64,7 @@ int Program::ConvertFullyConnected(hal::Operation* operation) {
     auto reshape_name = GetOperatorName(input_operand);
     auto reshape_op = std::make_shared<ge::op::Reshape>(reshape_name);
     auto shape_operator = AddInt32ConstantOperator(
-        std::vector<int32_t>({batch_size, input_size}));
+        std::vector<int32_t>({static_cast<int32_t>(batch_size), input_size}));
     SET_INPUT(reshape_op, x, input_operator);
     SET_INPUT(reshape_op, shape, shape_operator);
     input_operator = MAP_OUTPUT(reshape_op, y, input_operand);
